+++ packages/google_mlkit_genai_prompt/ios/Classes/GoogleMlKitGenaiPromptPlugin.h
@@ -1,0 +1,4 @@
+#import <Flutter/Flutter.h>
+
+@interface GoogleMlKitGenaiPromptPlugin : NSObject<FlutterPlugin>
+@end