+++ packages/google_mlkit_genai_proofreading/android/src/main/java/com/google_mlkit_genai_proofreading/GoogleMlKitGenaiProofreadingPlugin.java
@@ -1,0 +1,22 @@
+package com.google_mlkit_genai_proofreading;
+
+import androidx.annotation.NonNull;
+
+import io.flutter.embedding.engine.plugins.FlutterPlugin;
+import io.flutter.plugin.common.MethodChannel;
+
+public class GoogleMlKitGenaiProofreadingPlugin implements FlutterPlugin {
+    private MethodChannel channel;
+    private static final String channelName = "google_mlkit_genai_proofreading";
+
+    @Override
+    public void onAttachedToEngine(@NonNull FlutterPluginBinding flutterPluginBinding) {
+        channel = new MethodChannel(flutterPluginBinding.getBinaryMessenger(), channelName);
+        channel.setMethodCallHandler(new Proofreader(flutterPluginBinding.getApplicationContext()));
+    }
+
+    @Override
+    public void onDetachedFromEngine(@NonNull FlutterPluginBinding binding) {
+        channel.setMethodCallHandler(null);
+    }
+}