+++ packages/google_mlkit_genai_rewriting/lib/google_mlkit_genai_rewriting.dart
@@ -1,0 +1,4 @@
+export 'package:google_mlkit_commons/google_mlkit_commons.dart';
+
+export 'src/rewriter.dart'
+    show Rewriter, RewritingOutputType, RewritingLanguage;