+++ packages/google_mlkit_genai_summarization/ios/Classes/GoogleMlKitGenaiSummarizationPlugin.h
@@ -1,0 +1,4 @@
+#import <Flutter/Flutter.h>
+
+@interface GoogleMlKitGenaiSummarizationPlugin : NSObject<FlutterPlugin>
+@end