+++ packages/google_mlkit_genai_summarization/lib/google_mlkit_genai_summarization.dart
@@ -1,0 +1,8 @@
+export 'package:google_mlkit_commons/google_mlkit_commons.dart';
+
+export 'src/summarizer.dart'
+    show
+        Summarizer,
+        SummarizationInputType,
+        SummarizationOutputType,
+        SummarizationLanguage;